♻️ Replace TwoPaneBookDetail with unified WideBookDetail layout
